Consolidate job and invoice pages
