+++ lib/pages/customer_notifications_page.dart
@@ -9,12 +9,6 @@ class CustomerNotificationsPage extends StatelessWidget {
 
   const CustomerNotificationsPage({super.key, required this.userId});
 
-  String _formatDate(Timestamp? ts) {
-    if (ts == null) return '';
-    final dt = ts.toDate().toLocal();
-    return DateFormat('MM/dd h:mm a').format(dt);
-  }
-
   @override
   Widget build(BuildContext context) {
     final stream = FirebaseFirestore.instance
@@ -58,7 +52,7 @@ class CustomerNotificationsPage extends StatelessWidget {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     if (body.isNotEmpty) Text(body),
-                    Text(_formatDate(ts)),
+                    Text(formatDate(ts)),
                   ],
                 ),
                 onTap: () {
