+++ lib/pages/mechanic_radius_history_page.dart
@@ -2,18 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:intl/intl.dart';
 
+import "../utils.dart";
 /// Displays the history of radius changes for a mechanic.
 class MechanicRadiusHistoryPage extends StatelessWidget {
   final String mechanicId;
 
   const MechanicRadiusHistoryPage({super.key, required this.mechanicId});
 
-  String _formatDate(Timestamp? ts) {
-    if (ts == null) return '';
-    final dt = ts.toDate().toLocal();
-    return DateFormat('MMMM d, yyyy h:mm a').format(dt);
-  }
-
+  String formatDate(Timestamp? ts) {
   @override
   Widget build(BuildContext context) {
     final stream = FirebaseFirestore.instance
@@ -46,7 +42,7 @@ class MechanicRadiusHistoryPage extends StatelessWidget {
               return ListTile(
                 leading: const Icon(Icons.radio_button_checked),
                 title: Text('Radius: ${radius.toStringAsFixed(0)} miles'),
-                subtitle: Text(_formatDate(ts)),
+                subtitle: Text(formatDate(ts)),
               );
             },
           );
